+++ mon-pix/tests/acceptance/start-campaigns-workflow-test.js
@@ -1199,61 +1199,4 @@ module('Acceptance | Campaigns | Start Campaigns workflow', function (hooks) {
       });
     });
   });
-
-  module('Organizations access workflow when user is not logged in', function () {
-    let combinedCourse;
-
-    module('when user is not logged in', function (hooks) {
-      hooks.beforeEach(function () {
-        combinedCourse = server.create('combined-course', { id: 'COMBINIX1', organizationId: 1 });
-        server.create('verified-code', { id: 'COMBINIX1', type: 'combined-course', combinedCourse });
-        server.create('organization-to-join', { id: 1, isRestricted: false, code: combinedCourse.code });
-      });
-
-      module('when user tries to access join url', function () {
-        module('when user tries to access course as anonymous', function () {
-          test('should redirect to connection page', async function (assert) {
-            await visit(`organizations/${combinedCourse.code}/rejoindre/anonyme`);
-            assert.strictEqual(currentURL(), `/connexion`);
-          });
-        });
-        module('when user tries to access student-sco', function () {
-          test('should redirect to connection page', async function (assert) {
-            await visit(`organizations/${combinedCourse.code}/rejoindre/identification`);
-            assert.strictEqual(currentURL(), `/connexion`);
-          });
-        });
-        module('when user tries to access sco-mediacentre', function () {
-          test('should redirect to connection page', async function (assert) {
-            await visit(`organizations/${combinedCourse.code}/rejoindre/mediacentre`);
-            assert.strictEqual(currentURL(), `/connexion`);
-          });
-        });
-      });
-      module('when user tries to access invited url', function () {
-        test('should redirect to connection page', async function (assert) {
-          await visit(`organizations/${combinedCourse.code}/prescrit`);
-          assert.strictEqual(currentURL(), `/connexion`);
-        });
-        module('when user tries to access reconciliation page directly', function () {
-          test('should redirect to connection page', async function (assert) {
-            await visit(`organizations/${combinedCourse.code}/invited/enregistrement`);
-            assert.strictEqual(currentURL(), `/connexion`);
-          });
-        });
-        module('when user tries to access student-sco as invited', function () {
-          test('should redirect to connection page', async function (assert) {
-            await visit(`organizations/${combinedCourse.code}/invited/eleve`);
-            assert.strictEqual(currentURL(), `/connexion`);
-          });
-        });
-        module('when user tries to access sco-mediacentre as invited', function () {
-          test('should redirect to connection page', async function (assert) {
-            await visit(`organizations/${combinedCourse.code}/invited/etudiant`);
-            assert.strictEqual(currentURL(), `/connexion`);
-          });
-        });
-      });
-    });
-  });
 });
